refactor: split model loader from model definitions
